release(desktop): Release v0.6.0

+++ packages/internal/store/src/collection/store.ts
@@ -3,7 +3,9 @@ import type { CollectionSchema } from "@follow/database/schemas/types"
 import { CollectionService } from "@follow/database/services/collection"
 
 import { apiClient } from "../context"
-import type { Hydratable } from "../internal/base"
+import { getEntry } from "../entry/getter"
+import { invalidateEntriesQuery } from "../entry/hooks"
+import type { Hydratable, Resetable } from "../internal/base"
 import { createTransaction, createZustandStore } from "../internal/helper"
 
 interface CollectionState {
@@ -18,26 +20,22 @@ export const useCollectionStore = createZustandStore<CollectionState>("collectio
   () => defaultState,
 )
 
-// const get = useCollectionStore.getState
+const get = useCollectionStore.getState
 const set = useCollectionStore.setState
 
 class CollectionSyncService {
-  async starEntry({
-    entryId,
-    feedId,
-    view,
-  }: {
-    entryId: string
-    feedId: string
-    view: FeedViewType
-  }) {
+  async starEntry({ entryId, view }: { entryId: string; view: FeedViewType }) {
+    const entry = getEntry(entryId)
+    if (!entry) {
+      return
+    }
     const tx = createTransaction()
     tx.store(async () => {
       await collectionActions.upsertMany([
         {
           createdAt: new Date().toISOString(),
           entryId,
-          feedId,
+          feedId: entry.feedId,
           view,
         },
       ])
@@ -55,6 +53,8 @@ class CollectionSyncService {
     })
 
     await tx.run()
+
+    invalidateEntriesQuery({ collection: true })
   }
 
   async unstarEntry(entryId: string) {
@@ -78,20 +78,24 @@ class CollectionSyncService {
     })
 
     await tx.run()
+
+    invalidateEntriesQuery({ collection: true })
   }
 }
 
-class CollectionActions implements Hydratable {
+class CollectionActions implements Hydratable, Resetable {
   async hydrate() {
     const collections = await CollectionService.getCollectionAll()
     collectionActions.upsertManyInSession(collections)
   }
 
-  async upsertManyInSession(collections: CollectionSchema[]) {
-    const state = useCollectionStore.getState()
-    const nextCollections: CollectionState["collections"] = {
-      ...state.collections,
-    }
+  upsertManyInSession(collections: CollectionSchema[], options?: { reset?: boolean }) {
+    const state = get()
+    const nextCollections: CollectionState["collections"] = options?.reset
+      ? {}
+      : {
+          ...state.collections,
+        }
     collections.forEach((collection) => {
       if (!collection.entryId) return
       nextCollections[collection.entryId] = collection
@@ -102,42 +106,63 @@ class CollectionActions implements Hydratable {
     })
   }
 
-  async upsertMany(collections: CollectionSchema[]) {
+  async upsertMany(collections: CollectionSchema[], options?: { reset?: boolean }) {
     const tx = createTransaction()
     tx.store(() => {
-      this.upsertManyInSession(collections)
+      this.upsertManyInSession(collections, options)
     })
     tx.persist(() => {
-      return CollectionService.upsertMany(collections)
+      return CollectionService.upsertMany(collections, options)
     })
-    tx.run()
+    await tx.run()
   }
 
-  async deleteInSession(entryId: string) {
+  deleteInSession(entryId: string | string[]) {
+    const normalizedEntryId = Array.isArray(entryId) ? entryId : [entryId]
+
     const state = useCollectionStore.getState()
     const nextCollections: CollectionState["collections"] = {
       ...state.collections,
     }
-    delete nextCollections[entryId]
+
+    normalizedEntryId.forEach((id) => {
+      delete nextCollections[id]
+    })
     set({
       ...state,
       collections: nextCollections,
     })
   }
 
-  async delete(entryId: string) {
+  async delete(entryId: string | string[]) {
+    const entryIdsInCollection = new Set(Object.keys(get().collections))
+    const normalizedEntryId = (Array.isArray(entryId) ? entryId : [entryId]).filter((id) =>
+      entryIdsInCollection.has(id),
+    )
+
+    if (normalizedEntryId.length === 0) return
+
     const tx = createTransaction()
     tx.store(() => {
       this.deleteInSession(entryId)
     })
     tx.persist(() => {
-      return CollectionService.delete(entryId)
+      return CollectionService.deleteMany(normalizedEntryId)
     })
     tx.run()
   }
 
-  reset() {
-    set(defaultState)
+  async reset() {
+    const tx = createTransaction()
+    tx.store(() => {
+      set(defaultState)
+    })
+
+    tx.persist(() => {
+      return CollectionService.reset()
+    })
+
+    await tx.run()
   }
 }
